マルチスレッドに変更
