--- chari-spot/camera-system/pyproject.toml
@@ -1,17 +1,0 @@
-[tool.poetry]
-name = "camera-system"
-version = "0.1.0"
-description = ""
-authors = ["ReiMatsui <rei0203@icloud.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.11"
-opencv-python = "^4.11.0.86"
-loguru = "^0.7.3"
-ultralytics = "^8.3.113"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
