+++ chari-spot/camera_system/src/utils/camera_manager.py
@@ -1,0 +1,34 @@
+import cv2
+import numpy
+from loguru import logger
+
+class CameraManager:
+    """
+    カメラ映像を取得するクラス
+    """
+    def __init__(self, camera_no: int = 0):
+        self.capture = cv2.VideoCapture(camera_no)
+        
+        # 画質を取得し低めに設定する
+        self.width = int(self.capture.get(cv2.CAP_PROP_FRAME_WIDTH))
+        self.height = int(self.capture.get(cv2.CAP_PROP_FRAME_HEIGHT))
+            
+        # FPSを取得
+        self.fps = self.capture.get(cv2.CAP_PROP_FPS)
+        logger.info(f"カメラのFPS: {self.fps}")
+
+    def get_frames(self):
+        self.ret, self.frame = self.capture.read()
+        
+        if self.ret:
+            return cv2.flip(self.frame, 1)
+        return None, None
+    
+    def imshow(self, window_name: str, image: numpy.ndarray):
+        try:
+            cv2.imshow(window_name, image)
+        except Exception as e:
+            logger.error(f"画像表示/保存中のエラー: {e}")  
+            
+    def release(self):
+        self.capture.release()
